Fix implicit signed conversion and sign comparison
Compiled with clang's -Wpointer-sign -Wsign-conversion -Wsign-compare
-Wsign-promo flags, and fixed all reported implicit signed conversions
and comparison of signed and unsigned values. Most of these where
resolved by adding explicit type casts, but manually reviewed context of
this conversion to ensure overflow/underflow conditions where detected
in error cases.

In buffer module added BUFFER_MAX_SIZE constant set to 1GiB, this is to
avoid ssize_t overflow on 32 bit platforms.

+++ src/http.c
@@ -40,8 +40,8 @@ static const char http_503[] =
     "Backend not available";
 
 static int parse_http_header(const char *, size_t, char **);
-static int get_header(const char *, const char *, int, char **);
-static int next_header(const char **, int *);
+static int get_header(const char *, const char *, size_t, char **);
+static size_t next_header(const char **, size_t *);
 
 static const struct Protocol http_protocol_st = {
     .name = "http",
@@ -92,8 +92,8 @@ parse_http_header(const char* data, size_t data_len, char **hostname) {
 }
 
 static int
-get_header(const char *header, const char *data, int data_len, char **value) {
-    int len, header_len;
+get_header(const char *header, const char *data, size_t data_len, char **value) {
+    size_t len, header_len;
 
     header_len = strlen(header);
 
@@ -122,9 +122,9 @@ get_header(const char *header, const char *data, int data_len, char **value) {
     return -2;
 }
 
-static int
-next_header(const char **data, int *len) {
-    int header_len;
+static size_t
+next_header(const char **data, size_t *len) {
+    size_t header_len;
 
     /* perhaps we can optimize this to reuse the value of header_len, rather
      * than scanning twice.
